Attempt to fix tests
