+++ scriptworker/artifacts.py
@@ -1,3 +1,8 @@
+"""Scriptworker artifact-related operations.
+
+Importing this script updates the mimetypes database. This maps some known extensions to text/plain for a better storage
+in S3.
+"""
 import aiohttp
 import arrow
 import asyncio
@@ -17,77 +22,22 @@ from scriptworker.utils import download_file, filepaths_in_dir, raise_future_exc
 log = logging.getLogger(__name__)
 
 
-GZIP_SUPPORTED_CONTENT_TYPE = ('text/plain', 'application/json', 'text/html', 'application/xml')
-EXTENSIONS_TO_FORCE_TO_PLAIN_TEXT = ('.asc', '.log')
+_GZIP_SUPPORTED_CONTENT_TYPE = ('text/plain', 'application/json', 'text/html', 'application/xml')
+_EXTENSIONS_TO_FORCE_TO_PLAIN_TEXT = ('.asc', '.log')
 
 
-# create_artifact {{{1
-async def create_artifact(context, path, target_path, content_type, content_encoding, storage_type='s3', expires=None):
-    """Create an artifact and upload it.
+def _force_mimetypes_to_plain_text():
+    """Populate/Update the mime types database with supported extensions that we want to map to 'text/plain'.
 
-    This should support s3 and azure out of the box; we'll need some tweaking
-    if we want to support redirect/error artifacts.
-
-    Args:
-        context (scriptworker.context.Context): the scriptworker context.
-        path (str): the path of the file to upload.
-        target_path (str):
-        content_type (str): Content type (MIME type) of the artifact. Values can be found via
-            scriptworker.artifacts.guess_content_type_and_encoding()
-        content_encoding (str): Encoding (per mimetypes' library) of the artifact. None is for no encoding. Values can
-            be found via scriptworker.artifacts.guess_content_type_and_encoding()
-        storage_type (str, optional): the taskcluster storage type to use.
-            Defaults to 's3'
-        expires (str, optional): datestring of when the artifact expires.
-            Defaults to None.
-
-    Raises:
-        ScriptWorkerRetryException: on failure.
+    These extensions can then be open natively read by browsers once they're uploaded on S3. It doesn't affect artifacts
+    once they're downloaded from S3.
     """
-    payload = {
-        "storageType": storage_type,
-        "expires": expires or get_expiration_arrow(context).isoformat(),
-        "contentType": content_type,
-    }
-    args = [get_task_id(context.claim_task), get_run_id(context.claim_task),
-            target_path, payload]
-
-    tc_response = await context.temp_queue.createArtifact(*args)
-    skip_auto_headers = [aiohttp.hdrs.CONTENT_TYPE]
-    log.info("uploading {path} to {url}...".format(path=path, url=tc_response['putUrl']))
-    with open(path, "rb") as fh:
-        with aiohttp.Timeout(context.config['artifact_upload_timeout']):
-            async with context.session.put(
-                tc_response['putUrl'], data=fh, headers=craft_artifact_put_headers(content_type, content_encoding),
-                skip_auto_headers=skip_auto_headers, compress=False
-            ) as resp:
-                log.info(resp.status)
-                response_text = await resp.text()
-                log.info(response_text)
-                if resp.status not in (200, 204):
-                    raise ScriptWorkerRetryException(
-                        "Bad status {}".format(resp.status),
-                    )
+    for extension in _EXTENSIONS_TO_FORCE_TO_PLAIN_TEXT:
+        mimetypes.add_type('text/plain', extension)
+        log.debug('Extension "{}" forced to text/plain'.format(extension))
 
 
-# retry_create_artifact {{{1
-async def retry_create_artifact(*args, **kwargs):
-    """Retry create_artifact() calls.
-
-    Args:
-        *args: the args to pass on to create_artifact
-        **kwargs: the args to pass on to create_artifact
-    """
-    await retry_async(
-        create_artifact,
-        retry_exceptions=(
-            ScriptWorkerRetryException,
-            aiohttp.errors.DisconnectedError,
-            aiohttp.errors.ClientError
-        ),
-        args=args,
-        kwargs=kwargs
-    )
+_force_mimetypes_to_plain_text()
 
 
 # upload_artifacts {{{1
@@ -131,6 +81,133 @@ async def upload_artifacts(context):
             )
         )
     await raise_future_exceptions(tasks)
+
+
+def compress_artifact_if_supported(artifact_path):
+    """Compress artifacts with GZip if they're known to be supported.
+
+    This replaces the artifact given by a gzip binary. Nothing happens if the artifact is already zip'd
+
+    Args:
+        artifact_path (str): the path to compress
+
+    Returns:
+        content_type, content_encoding (tuple):  Type and encoding of the file. Encoding equals 'gzip' if compressed.
+    """
+    content_type, encoding = guess_content_type_and_encoding(artifact_path)
+    log.debug('"{}" is encoded with "{}" and has mime/type "{}"'.format(artifact_path, encoding, content_type))
+
+    if encoding is None and content_type in _GZIP_SUPPORTED_CONTENT_TYPE:
+        log.debug('"{}" can be gzip\'d. Compressing...'.format(artifact_path))
+        with open(artifact_path, 'rb') as f_in:
+            text_content = f_in.read()
+
+        with gzip.open(artifact_path, 'wb') as f_out:
+            f_out.write(text_content)
+
+        encoding = 'gzip'
+        log.info('Compressed "{}" with gzip'.format(artifact_path))
+    else:
+        log.debug('"{}" is not supported for compression.'.format(artifact_path))
+
+    return content_type, encoding
+
+
+def guess_content_type_and_encoding(path):
+    """Guess the content type of a path, using ``mimetypes``.
+
+    Falls back to "application/binary" if no content type is found.
+
+    Args:
+        path (str): the path to guess the mimetype of
+
+    Returns:
+        str: the content type of the file
+    """
+    content_type, encoding = mimetypes.guess_type(path)
+    content_type = content_type or "application/binary"
+    return content_type, encoding
+
+
+# retry_create_artifact {{{1
+async def retry_create_artifact(*args, **kwargs):
+    """Retry create_artifact() calls.
+
+    Args:
+        *args: the args to pass on to create_artifact
+        **kwargs: the args to pass on to create_artifact
+    """
+    await retry_async(
+        create_artifact,
+        retry_exceptions=(
+            ScriptWorkerRetryException,
+            aiohttp.errors.DisconnectedError,
+            aiohttp.errors.ClientError
+        ),
+        args=args,
+        kwargs=kwargs
+    )
+
+
+# create_artifact {{{1
+async def create_artifact(context, path, target_path, content_type, content_encoding, storage_type='s3', expires=None):
+    """Create an artifact and upload it.
+
+    This should support s3 and azure out of the box; we'll need some tweaking
+    if we want to support redirect/error artifacts.
+
+    Args:
+        context (scriptworker.context.Context): the scriptworker context.
+        path (str): the path of the file to upload.
+        target_path (str):
+        content_type (str): Content type (MIME type) of the artifact. Values can be found via
+            scriptworker.artifacts.guess_content_type_and_encoding()
+        content_encoding (str): Encoding (per mimetypes' library) of the artifact. None is for no encoding. Values can
+            be found via scriptworker.artifacts.guess_content_type_and_encoding()
+        storage_type (str, optional): the taskcluster storage type to use.
+            Defaults to 's3'
+        expires (str, optional): datestring of when the artifact expires.
+            Defaults to None.
+
+    Raises:
+        ScriptWorkerRetryException: on failure.
+    """
+    payload = {
+        "storageType": storage_type,
+        "expires": expires or get_expiration_arrow(context).isoformat(),
+        "contentType": content_type,
+    }
+    args = [get_task_id(context.claim_task), get_run_id(context.claim_task),
+            target_path, payload]
+
+    tc_response = await context.temp_queue.createArtifact(*args)
+    skip_auto_headers = [aiohttp.hdrs.CONTENT_TYPE]
+    log.info("uploading {path} to {url}...".format(path=path, url=tc_response['putUrl']))
+    with open(path, "rb") as fh:
+        with aiohttp.Timeout(context.config['artifact_upload_timeout']):
+            async with context.session.put(
+                tc_response['putUrl'], data=fh, headers=_craft_artifact_put_headers(content_type, content_encoding),
+                skip_auto_headers=skip_auto_headers, compress=False
+            ) as resp:
+                log.info(resp.status)
+                response_text = await resp.text()
+                log.info(response_text)
+                if resp.status not in (200, 204):
+                    raise ScriptWorkerRetryException(
+                        "Bad status {}".format(resp.status),
+                    )
+
+
+def _craft_artifact_put_headers(content_type, encoding=None):
+    log.debug('{} {}'.format(content_type, encoding))
+    headers = {
+        aiohttp.hdrs.CONTENT_TYPE: content_type,
+    }
+
+    if encoding is not None:
+        headers[aiohttp.hdrs.CONTENT_ENCODING] = encoding
+
+    return headers
 
 
 # get_artifact_url {{{1
@@ -224,71 +301,3 @@ async def download_artifacts(context, file_urls, parent_dir=None, session=None,
 
     await raise_future_exceptions(tasks)
     return files
-
-
-def compress_artifact_if_supported(artifact_path):
-    """ Compress artifacts with GZip if they're known to be supported.
-
-    This replaces the artifact given by a gzip binary. Nothing happens if the artifact is already zip'd
-
-    Args:
-        artifact_path (str): the path to compress
-
-    Returns:
-        content_type, content_encoding (tuple):  Type and encoding of the file. Encoding equals 'gzip' if compressed.
-    """
-    content_type, encoding = guess_content_type_and_encoding(artifact_path)
-    log.debug('"{}" is encoded with "{}" and has mime/type "{}"'.format(artifact_path, encoding, content_type))
-
-    if content_type in GZIP_SUPPORTED_CONTENT_TYPE:
-        log.debug('"{}" can be gzip\'d. Compressing...'.format(artifact_path))
-        with open(artifact_path, 'rb') as f_in:
-            text_content = f_in.read()
-
-        with gzip.open(artifact_path, 'wb') as f_out:
-            f_out.write(text_content)
-        encoding = 'gzip'
-        log.info('Compressed "{}" with gzip'.format(artifact_path))
-    else:
-        log.debug('"{}" is not supported for compression.'.format(artifact_path))
-
-    return content_type, encoding
-
-
-def guess_content_type_and_encoding(path):
-    """Guess the content type of a path, using ``mimetypes``.
-
-    Args:
-        path (str): the path to guess the mimetype of
-
-    Returns:
-        str: the content type of the file
-    """
-    content_type, encoding = mimetypes.guess_type(path)
-    content_type = content_type or "application/binary"
-    return content_type, encoding
-
-
-def craft_artifact_put_headers(content_type, encoding=None):
-    log.debug('{} {}'.format(content_type, encoding))
-    headers = {
-        aiohttp.hdrs.CONTENT_TYPE: content_type,
-    }
-
-    if encoding is not None:
-        headers[aiohttp.hdrs.CONTENT_ENCODING] = encoding
-
-    return headers
-
-
-def _update_mimetypes_to_plain_text():
-    """Populate/Update the mime types database with supported extensions that we want to map to 'text/plain'
-
-    These extensions can then be open natively read by browsers once they're uploaded on S3.
-    """
-    for extension in EXTENSIONS_TO_FORCE_TO_PLAIN_TEXT:
-        mimetypes.add_type('text/plain', extension)
-        log.debug('Extension "{}" forced to text/plain'.format(extension))
-
-
-_update_mimetypes_to_plain_text()
