+++ scriptworker/test/test_task.py
@@ -2,14 +2,12 @@
 # coding=utf-8
 """Test scriptworker.task
 """
-import arrow
 import asyncio
 import glob
 import mock
 import os
 import pprint
 import pytest
-from scriptworker.exceptions import ScriptWorkerRetryException
 import scriptworker.task as task
 import scriptworker.log as log
 import sys
@@ -17,7 +15,7 @@ import taskcluster.exceptions
 import taskcluster.async
 import time
 from . import event_loop, fake_session, fake_session_500, rw_context, successful_queue, \
-    touch, unsuccessful_queue, read
+    unsuccessful_queue, read
 
 assert event_loop, rw_context  # silence flake8
 assert fake_session, fake_session_500  # silence flake8
@@ -25,14 +23,6 @@ assert successful_queue, unsuccessful_queue  # silence flake8
 
 # constants helpers and fixtures {{{1
 TIMEOUT_SCRIPT = os.path.join(os.path.dirname(__file__), "data", "long_running.py")
-
-mimetypes = {
-    "/foo/bar/test.txt": "text/plain",
-    "/tmp/blah.tgz": "application/x-tar",
-    "~/Firefox.dmg": "application/x-apple-diskimage",
-    "/foo/bar/blah.log": "text/plain",
-    "/totally/unknown": "application/binary",
-}
 
 
 @pytest.yield_fixture(scope='function')
@@ -54,25 +44,6 @@ def context(rw_context):
 @pytest.mark.parametrize("one,two,expected", ((1, 2, 2), (4, 2, 4)))
 def test_worst_level(one, two, expected):
     assert task.worst_level(one, two) == expected
-
-
-# get_expiration_arrow {{{1
-def test_expiration_arrow(context):
-    now = arrow.utcnow()
-
-    # make sure time differences don't screw up the test
-    with mock.patch.object(arrow, 'utcnow') as p:
-        p.return_value = now
-        expiration = task.get_expiration_arrow(context)
-        diff = expiration.timestamp - now.timestamp
-        assert diff == 3600
-
-
-# guess_content_type {{{1
-@pytest.mark.parametrize("mimetypes", [(k, v) for k, v in sorted(mimetypes.items())])
-def test_guess_content_type(mimetypes):
-    path, mimetype = mimetypes
-    assert task.guess_content_type(path) == mimetype
 
 
 # get_decision_task_id {{{1
@@ -178,62 +149,6 @@ async def test_reclaim_task_mock(context, mocker, event_loop):
     await task.reclaim_task(context, context.task)
 
 
-# upload_artifacts {{{1
-def test_upload_artifacts(context, event_loop):
-    args = []
-    os.makedirs(os.path.join(context.config['artifact_dir'], 'public'))
-    paths = [
-        os.path.join(context.config['artifact_dir'], 'one'),
-        os.path.join(context.config['artifact_dir'], 'public/two'),
-    ]
-    for path in paths:
-        touch(path)
-
-    async def foo(_, path, **kwargs):
-        args.append(path)
-
-    with mock.patch('scriptworker.task.create_artifact', new=foo):
-        event_loop.run_until_complete(
-            task.upload_artifacts(context)
-        )
-
-    assert sorted(args) == sorted(paths)
-
-
-# create_artifact {{{1
-def test_create_artifact(context, fake_session, successful_queue, event_loop):
-    path = os.path.join(context.config['artifact_dir'], "one.txt")
-    touch(path)
-    context.session = fake_session
-    expires = arrow.utcnow().isoformat()
-    context.temp_queue = successful_queue
-    event_loop.run_until_complete(
-        task.create_artifact(context, path, "public/env/one.txt", expires=expires)
-    )
-    assert successful_queue.info == [
-        "createArtifact", ('taskId', 'runId', "public/env/one.txt", {
-            "storageType": "s3",
-            "expires": expires,
-            "contentType": "text/plain",
-        }), {}
-    ]
-    context.session.close()
-
-
-def test_create_artifact_retry(context, fake_session_500, successful_queue,
-                               event_loop):
-    path = os.path.join(context.config['artifact_dir'], "one.log")
-    touch(path)
-    context.session = fake_session_500
-    expires = arrow.utcnow().isoformat()
-    with pytest.raises(ScriptWorkerRetryException):
-        context.temp_queue = successful_queue
-        event_loop.run_until_complete(
-            task.create_artifact(context, path, "public/env/one.log", expires=expires)
-        )
-    context.session.close()
-
-
 # max_timeout {{{1
 def test_max_timeout_noop(context):
     with mock.patch.object(task.log, 'debug') as p:
@@ -260,43 +175,3 @@ def test_max_timeout(context, event_loop):
         print("Checking {}...".format(path))
         assert files[path] == (time.ctime(os.path.getmtime(path)), os.stat(path).st_size)
     assert len(files.keys()) == 6
-
-
-# get_artifact_url {{{1
-def test_get_artifact_url():
-
-    def makeRoute(*args, **kwargs):
-        return "rel/path"
-
-    context = mock.MagicMock()
-    context.queue = mock.MagicMock()
-    context.queue.options = {'baseUrl': 'https://netloc/'}
-    context.queue.makeRoute = makeRoute
-    assert task.get_artifact_url(context, "x", "y") == "https://netloc/v1/rel/path"
-
-
-# download_artifacts {{{1
-def test_download_artifacts(context, event_loop):
-    urls = []
-    paths = []
-
-    expected_urls = [
-        "https://queue.taskcluster.net/v1/task/dependency1/artifacts/foo/bar",
-        "https://queue.taskcluster.net/v1/task/dependency2/artifacts/baz",
-    ]
-    expected_paths = [
-        os.path.join(context.config['work_dir'], "foo", "bar"),
-        os.path.join(context.config['work_dir'], "baz"),
-    ]
-
-    async def foo(_, url, path, **kwargs):
-        urls.append(url)
-        paths.append(path)
-
-    result = event_loop.run_until_complete(
-        task.download_artifacts(context, expected_urls, download_func=foo)
-    )
-
-    assert sorted(result) == sorted(expected_paths)
-    assert sorted(paths) == sorted(expected_paths)
-    assert sorted(urls) == sorted(expected_urls)
